refactor(rust): move transfer handlers to separate modules
